Feat: Add new recipe
Created endpoint for chef's use to create new recipes from the ingredients in the database. It first creates a new ManufactoredIngredient and commit the changes. Then it adds the used ingredient ids and their quantity to the database if not already present, finally adds them to the recipe along the newly created ManufactoredIngredient id. Added createrecipe.html, more on README on why I didn't use wtforms.